blas.buildlink3.mk: remove TODOs
Those were vague plans for the future and do not relate to
missing functionality. It is obvious that we will want to
support more implementations in future, as they enter pkgsrc
proper. The idea about pkg-config files is similar. Once
stuff provides them and other stuff can use them, we will
add a variable to locate them.