residualvm-git: import residualvm-0.4.0 (git) as wip/residualvm-git
Residual is a ScummVM sub-project to play LucasArts' LUA-based 3D
adventures, such as Grim Fandango. Residual is an OpenGL program,
and requires a 3D card with working OpenGL support for optimum
playability.

A software renderer is also included for machines without hardware
OpenGL.

This package is built from the latest git snapshots.