Import pylru as wip/py-pylru version 1.2.0
Pylru implements a true LRU cache along with several support classes. The cache
is efficient and written in pure Python. It works with Python 2.6+ including
the 3.x series. Basic operations (lookup, insert, delete) all run in a constant
amount of time. Pylru provides a cache class with a simple dict interface. It
also provides classes to wrap any object that has a dict interface with a
cache. Both write-through and write-back semantics are supported. Pylru also
provides classes to wrap functions in a similar way, including a function
decorator.