flnews-devel: Update to snapshot 0.17pre18
The Format=flowed decoder now supports a mode that adds an empty line
if a paragraph ends with an empty line. This new mode can be enabled
with the entry flowed_insert_crlf in configfile or with the experimantal
MIME parameter InsLine=yes in the Content-Type headerfield.

Paragraphs that end with an empty line are useful e.g. if the sender wants
to allow rewrap (to smaller width) for single lines. RFC 3676 does not
allow paragraphs containing only a single line.