py-git: Update to 3.0.5
pkgsrc changes:
 - Extend DESCR based on contents available on HOMEPAGE

Changes:
3.0.5 - Bugfixes
=============================================
see the following for details:
https://github.com/gitpython-developers/gitpython/milestone/32?closed=1

3.0.4 - Bugfixes
=============================================
see the following for details:
https://github.com/gitpython-developers/gitpython/milestone/31?closed=1

3.0.3 - Bugfixes
=============================================
see the following for (most) details:
https://github.com/gitpython-developers/gitpython/milestone/30?closed=1

3.0.2 - Bugfixes
=============================================
* fixes an issue with installation

3.0.1 - Bugfixes and performance improvements
=============================================
* Fix a `performance regression <https://github.com/gitpython-developers/GitPython/issues/906>`_ which could make certain workloads 50% slower
* Add `currently_rebasing_on` method on `Repo`, see `the PR <https://github.com/gitpython-developers/GitPython/pull/903/files#diff-c276fc3c4df38382ec884e59657b869dR1065>`_
* Fix incorrect `requirements.txt` which could lead to broken installations, see this `issue <https://github.com/gitpython-developers/GitPython/issues/908>`_ for details.

3.0.0 - Remove Python 2 support
===============================
Motivation for this is a patch which improves unicode handling when dealing with filesystem paths.
Python 2 compatibility was introduced to deal with differences, and I thought it would be a good idea
to 'just' drop support right now, mere 5 months away from the official maintenance stop of python 2.7.

The underlying motivation clearly is my anger when thinking python and unicode, which was a hassle from the
start, at least in a codebase as old as GitPython, which totally doesn't handle encodings correctly in many cases.

Having migrated to using `Rust` exclusively for tooling, I still see that correct handling of encodings isn't entirely
trivial, but at least `Rust` makes clear what has to be done at compile time, allowing to write software that is pretty
much guaranteed to work once it compiles.

Again, my apologies if removing Python 2 support caused inconveniences, please see release 2.1.13 which returns it.

see the following for (most) details:
https://github.com/gitpython-developers/gitpython/milestone/27?closed=1

2.1.13 - Bring back Python 2.7 support
======================================
My apologies for any inconvenience this may have caused. Following semver, backward incompatible changes
will be introduced in a minor version.

2.1.12 - Bugfixes and Features
==============================
* Multi-value support and interface improvements for Git configuration. Thanks to A. Jesse Jiryu Davis.

2.1.11 - Bugfixes
=================
see the following for (most) details:
https://github.com/gitpython-developers/gitpython/milestone/26?closed=1

2.1.10 - Bugfixes
=================
see the following for (most) details:
https://github.com/gitpython-developers/gitpython/milestone/25?closed=1

2.1.9 - Dropping support for Python 2.6
=======================================
see the following for (most) details:
https://github.com/gitpython-developers/gitpython/milestone/24?closed=1

2.1.8 - bugfixes
====================================
see the following for (most) details:
https://github.com/gitpython-developers/gitpython/milestone/23?closed=1