cyclone: import cyclone-0.11.7 as wip/cyclone
Cyclone is an R7RS Scheme-to-C compiler that uses a variant of
Cheney on the MTA (http://www.pipeline.com/~hbaker1/CheneyMTA.html)
to implement full tail recursion, continuations, and generational
garbage collection.  In addition, the Cheney on the MTA concept
has been extended to allow execution of multiple native threads.
An on-the-fly garbage collector is used to manage the second-generation
heap and perform major collections without "stopping the world".

Cyclone is written entirely in the latest R7RS Scheme language
standard, and the intent is to support as much of that language as
possible.