fs2open-git: import fs2open (git snapshot) as fs2open-git
The Source Code Project was started roughly five years ago, when Volition
released the source code to the game known as "Freespace 2". This virtually
unknown game consistently won awards for being a great action space sim, but
never really caught on when it was released. The Source Code Project has worked,
and continues to work, on improving the graphics and gameplay of this
ten-year-old game.

This package installs the executable only. The original game CD, or
data files for the game are required to play the game.