Copy over www/grafana and apply Go 1.13 build fix.
The included copy of golang.org/x/xerrors is too old and does not
build with go 1.13. Make it use wip/go-xerrors instead.

This will be committed to pkgsrc itself after the branch, but
right now, it has no advantage (since the tree is still on Go 1.12)
and just adds noise during the freeze.