ats2: Update wip/ats2 to 0.4.0
0.4.0

This is the 51st release of ATS2, the successor of the
ATS programming language. The compiler for ATS2 is given
the name ATS/Positats, ATS2/Postiats or simply Postiats.

The official website for ATS is:

http://www.ats-lang.org

ATS-Postiats is hosted at github:

https://github.com/githwxi/ATS-Postiats

Major releases of ATS2 are available at:

https://sourceforge.net/projects/ats2-lang/

Major releases of external packages for ATS2 are available at:

https://sourceforge.net/projects/ats2-lang-contrib/

Here is a list of major additions and changes since the last release:

1. Improving testing code:
   doc/EXAMPLE/ATSLF,
   doc/EXAMPLE/INTRO, doc/EXAMPLE/TESTATS
   Kudos to Artyom Shalkhakov!
2. Introducing various keywords (e.g., atstflt, atsvtflt), which are
   used in the implementation of ATS3(ATS/Xanadu)

Please note that there are two variants of the package released: The
"gmp" variant depends on the GMP library while the "int" variant does
not. From this release on, ATS2-Postiats-0.4.0.tgz is just the same as
ATS2-Postiats-int-0.4.0.tgz. This choice is made due to the fact that
most users (especially new users) can readily get by without the extra
dependency on GMP.