fonttosfnt: update to 1.1.0.
Hot on the heels of the 2018 release, here's another fonttosfnt release to
get your new year going. See, it doesn't always take a decade to run a release
script and throw some tarballs at a server. Sometimes a year and a bit is
sufficient.

fonttosfnt is a commandline tool that wraps bitmap fonts into a truetype
wrapper. This release includes a few code changes that warrant the minor
version number bump, but for a summary the git shortlog is about as useful
as any summary I can come up with.

Alan Coopersmith (2):
      Update README for gitlab migration
      Update configure.ac bug URL for gitlab migration

Jouke Witteveen (3):
      Correct a string literal
      Mention .otb as an extension
      Accept a BDF font on stdin

Maya Rashish (1):
      Avoid undefined behaviour

Peng Wu (1):
      Round font size to integer value when reading bitmap font

Peter Hutterer (2):
      Mark vsprintf_alloc as printf-like function
      Bump to 1.1.0

Ryan Farley (3):
      check for freetype NULL atoms
      let freetype handle ISO-8859-1 mapping
      allow ISO-646.1991-IRV as well, adobe standard for bdf