"New package" - muparser, update to 2.2.6
Update muparser to 2.2.6
The version in pkgsrc is 200.2.2.2
The project moved from SourceForge to Github in 2016.

Package builds and installs without the patches applied
on the pkgsrc version.

Please review the need or, not of these patches.
Also, review the buildlink3.mk, as they differ.

I only needed this package as a build dependency and don't
know if testes are needed.