firefox69-pkgsrc2019Q4:essentially www/firefox (69.0.1) from pkgsrc-2019Q3.
It installs as firefox, not firefox69, so it can directly replace
www/firefox.

Because it failed to build for me when I tried to in the context of
2019-Q4, I randomly hacked on it until it did. I wanted to build it
because the normal firefox in 2019-Q4 is broken (about:logins, the saved
login page shows no logins), and you can't downgrade versions to
www/firefox68 (which did build).