New package, epson-inkjet-printer-escpr-1.7.7.
Asked for by Dmitrii Postolov on netbsd-users.

This software is a filter program used with the Common UNIX Printing
System (CUPS).  The software offers high quality printing with Seiko Epson
color ink jet printers.

This product can only be used with printers that support the Epson ESC/P-R
language.

Please see refer to the following for a list of supported printers.
http://download.ebz.epson.net/dsc/search/01/search/?OSC=LX