gh: Import gh-0.5.4 as wip/gh
gh is GitHub on the command line, and it's now available in beta.
It brings pull requests, issues, and other GitHub concepts to the
terminal next to where you are already working with git and your code.

Still does not build because all dependencies are not provided by the distfile,
please give a look to TODO for more information and/or let me know how we can
address that! (packaging all dependencies will probably be not fun!)