woof: Import version 1.0.1
Woof! is a continuation of Lee Killough's Doom source port MBF
targeted at modern systems.

MBF stands for "Marine's Best Friend" and is regarded by many as
the successor of the Boom source port by TeamTNT. It serves as the
code base for many of today's successful Doom source ports such as
PrBoom+ or The Eternity Engine. As the original engine was limited
to run only under MS-DOS, it has been ported to Windows by Team
Eternity under the name WinMBF in 2004. Woof! is developed based on
the WinMBF code with the aim to make MBF more widely available and
convenient to use on modern systems.

To achieve this goal, this source port is less strict regarding its
faithfulness to the original MBF. It is focused on quality-of-life
enhancements, bug fixes and compatibility improvements. However,
all changes have been introduced in good faith that they are in line
with the original author's intentions and even for the trained eye,
this source port should be hard to distinguish from the original MBF.