Featherpad, update to 0.13.0
The main change in this release is the support for syntax color
customization, with separate settings for light and dark backgrounds.
Also, the color value of whitespaces can be customized,
in case they're shown.

Other changes:

- Allow dash and some other characters in URLs.
- Considered extra cases of valid ampersands in HTML.
- Yet smarter Shift+Enter with alphabetical lists.
- Removed QCoreApplication::processEvents() to prevent a rare crash.
- Remember cursor's horizontal pixel position before Backspace/Enter and
restore it after Down/Up. (This feature was removed from Qt 5.14.1
with Backspace, probably due to a regression.)
- A few fixes.