tmux: Import tmux-3.1rc2 as wip/tmux
Tmux is a "terminal multiplexer", it enables a number of terminals
(or windows) to be accessed and controlled from a single terminal.
Tmux is intended to be a simple, modern, BSD-licensed alternative
to programs such as GNU screen.

This track release candidates of tmux and will be used to update
misc/tmux when the 3.1 stable release will be released.