Add glimpse. Instantly segfaults.
Glimpse is an open source image editor based on the GNU Image Manipulation
Program. The goal is to experiment with new ideas and expand the use of free
software.

The GNU Image Manipulation Program is primarily intended for long-standing
power users. We have rebranded and repackaged it for a wider audience, and
we intend to make it much easier to deploy across schools, university campuses
and workplaces. As the project develops we intend to focus on performance
improvements, user interface changes and incrementally improving the user
experience.