Remove LXQt, imported into main branch
Remove LXQt meta-pkg and remaining components (-admin, -sudo,
openssh-askpass, -policykit and -powermanagement)

These packages are now in the main pkgsrc branch.
Thanks to David H. Gutteridge for the awesome tweeks done
to vastly improve user experience.

Thanks also to Nikita for the early help and initial review
of all the DE components.

It has been an awesome learning experience working with both
of you and to bring LXQt to this stage. Thanks!