ajbsp: Import version 1.01
AJBSP is a simple nodes builder for modern DOOM source ports.
It can build standard DOOM nodes, GL-Nodes, and XNOD format nodes.
The code is based on the BSP code in Eureka DOOM Editor, which was
based on the code from glBSP but with significant changes.

AJBSP is a command-line tool. It can handle multiple wad files,
and modifies each file in-place. There is an option to backup each
file first. The output to the terminal is fairly terse, but greater
verbosity can be enabled. Generally all the maps in a wad will
processed, but this can be limited to a specific set.