openhexagon: reconfigure build
Bring all the tagged submodules into openhexagon directly. Tell
cmake know where the library/include files for each will be, then
configure and build. All relevant files will be found when they
are needed.

this means we don't need to worry about the lack of release
versions for the submodules as separate packages.