wip/libbacon: import libbacon-1.2.2.1
Libbacon is a collection of miscellaneous C functions that I felt were
generally useful enough that they should be made available in a library.
My hope is that many of them will someday be added to the existing C
libraries, such as libc, libm, etc.  ( A few, such as strlcpy(), and
strlcat() already have on some systems. )