lapack, lapacke: fix lapacke build, remove lapacke-targets stuff
This suddenly was broken, cmake failure in configure phase. Dunno why.
I followed the example of cblas and removed the offending install()
line. It resulted in extra .cmake files that seem unnecessary anyway.

We need to upstream that build hackery to avoid having to mess with
forces we (I) do not understand (cmake).