Add a draft package for clisp "current as of 7 jun 2020".
This is based on lang/clisp, with patches removed for now, and
only added those required to build on NetBSD/i386 9.0.

The old verison, 2.49, is from 2010, while this one is being worked
on.  The repository is moved to gitlab.com, but there's no releases
to point to (argh!), so we have to resort to referring to the
version via a tag.

Verified to build on NetBSD/i386 9.0; had to reduce the end of
mmapp'able memory due to ASLR.

Sadly, no change log in sight.