coccigrep: Import coccigrep-0.16nb1 as wip/coccigrep
coccigrep is a semantic grep for the C language based on coccinelle.
It can be used to find where a given structure is used in code
files. coccigrep depends on the spatch program which comes with
coccinelle.

This is intended to be used to be updated and to test wip/coccinelle.