plan9port: Add a patch to honors LDFLAGS for xshove
Previously it was not linked against libX11.
Now plan9port finally installs but there is still a lot of TODO to clean it up
and fix it (e.g. acme and probably all other graphical programs fail to run).

Please give a look to TODO and try it if you would like to help!