collectd: correct disk plugin implementation for NetBSD
Correct the disk plugin implementation for NetBSD in patch-src_disk.c.

The original author seems to have mis-understood the collect "derive" type. It doesn't mean the value provided is a derivative, but it's a gauge you're probably interested in the derivative of.
In fact, that makes the code a lot simpler.

See https://github.com/collectd/collectd/pull/3334 and https://github.com/collectd/collectd/pull/3379.