starfighter: update to starfighter-2.3
No changelog.

Announcement:

When someone I'm very close to played Project: Starfighter recently, we noticed
a major bugs in Project: Starfighter pertaining to gamepad control. Then I
noticed another bug with gamepad control. Whoops! So this release set out to
fix those bugs (one which caused the level end sequence to be messed up by use
of a joystick, and one which caused dpad-based movement to be at rocket speed)
as well as a typo that said person I'm close to noticed as well.

They also had a request for an accessibility feature allowing slower text
speed, and I thought, you know what? That's an excellent idea! So I added in an
option to the Options menu which adjust text speed between five options: Normal
(same as before), Fast, Instant, Slow, and Sloth. As I implemented these I also
noted that it would probably be a good idea to allow the player to press a
button to advance certain dialog boxes, so I added that as well: pressing the
fire buttons now advances to the next dialog box instantly rather than skipping
the entire cutscene as it did before (Escape still skips the entire cutscene).
This should be helpful for those who wish to use Slow or Sloth speed but not
have to wait inordinate periods of time through the cutscenes.

And that's it for this release! It's really a pretty simple one, but I decided
to make it version 2.3 instead of 2.2.2 because it adds features. Enjoy!