Import ncspot
ncurses Spotify client written in Rust using librespot. It is heavily inspired
by ncurses MPD clients, such as ncmpc.  My motivation was to provide a simple
and resource friendly alternative to the official client as well as to support
platforms that currently don't have a Spotify client, such as the *BSDs.

WARNING: This was an user request, I don't have a premium Spotify account.
The package fails to build, see TODO.
If someone else has any ideas to fix this, please let me know.