starfighter: update to starfighter-2.3.1
Announcement:

Time for another release!

A few things prompted this, but the start of it was when we wanted
to include a Japanese translation by maruhiro-ver0 (#8), as well
as to improve gamepad usage and support (#7). So, yeah, this version
includes said Japanese translation and improved gamepad support!

Along the way, we also made the following changes:

Adjusted in-mission dialogs to match the dialogs' timing (with one
exception, the first Kline encounter, so as not to mess with music
cues there).  Made it so that, during the Kline fight in Classic
difficulty, Kline's health goes to the health levels of respective
stages when stage changes. This is done to simulate the behavior
of the original, where Kline actually had four separate health bars
that didn't bleed into each other.  Fixed a bug in rendering of
Unicode text which allowed multi-byte Unicode characters to be
split into multiple parts. (#9)